[upd] delete function

+++ src/main/java/be/intecbrussel/repository/StorageRepository.java
@@ -1,11 +1,12 @@
 package be.intecbrussel.repository;
 
 import be.intecbrussel.config.RepoConfig;
+import be.intecbrussel.exceptions.StorageException;
+import be.intecbrussel.model.Key;
 import be.intecbrussel.model.Product;
 import be.intecbrussel.model.Storage;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.Query;
-
 
 public class StorageRepository {
 
@@ -14,12 +15,17 @@ public class StorageRepository {
             em.getTransaction().begin();
             em.persist(storage);
             em.getTransaction().commit();
+        } catch (Exception ex) {
+            throw new StorageException("Error creating storage", ex);
         }
     }
 
     public Storage readStorage(long id) {
         try (EntityManager em = RepoConfig.getEM()) {
-            return em.find(Storage.class, id);
+            Storage storage = em.find(Storage.class, id);
+            return storage;
+        } catch (Exception ex) {
+            throw new StorageException("Error reading storage", ex);
         }
     }
 
@@ -28,6 +34,18 @@ public class StorageRepository {
             Query query = em.createQuery("SELECT s FROM Storage s JOIN s.products p WHERE p.id = ?1");
             query.setParameter(1, product.getId());
             return (Storage) query.getSingleResult();
+        } catch (Exception ex) {
+            throw new StorageException("Error reading storage for product", ex);
+        }
+    }
+
+    public Storage readStorage(Key key) {
+        try (EntityManager em = RepoConfig.getEM()) {
+            Query query = em.createQuery("SELECT k.storage FROM Key k WHERE k.id = :keyId");
+            query.setParameter("keyId", key.getId());
+            return (Storage) query.getSingleResult();
+        } catch (Exception ex) {
+            throw new StorageException("Error reading storage for key", ex);
         }
     }
 
@@ -36,6 +54,8 @@ public class StorageRepository {
             em.getTransaction().begin();
             em.merge(storage);
             em.getTransaction().commit();
+        } catch (Exception ex) {
+            throw new StorageException("Error updating storage", ex);
         }
     }
 
@@ -43,9 +63,14 @@ public class StorageRepository {
         try (EntityManager em = RepoConfig.getEM()){
             em.getTransaction().begin();
             Storage dbStorage = em.find(Storage.class, id);
-            em.remove(dbStorage);
-            em.getTransaction().commit();
+            if (dbStorage != null) {
+                em.remove(dbStorage);
+                em.getTransaction().commit();
+            } else {
+                throw new StorageException("Storage with ID " + id + " not found for deletion.");
+            }
+        } catch (Exception ex) {
+            throw new StorageException("Error deleting storage", ex);
         }
     }
-
 }
